Allow to copy UI5 bundles from plugins folder to m2 repo
A new maven project setup is introduced. Building it will copy the
required UI5 jars from the eclipse installation to the local maven
repository.

The following properties need to be specified:
- sap.cloud.ui5.version - UI5 version
- eclipse.path path to eclipse installation

Th project itself can be build by invoking mvn clean install in the web
project. This is also how the project is build on Jenkins.

Change-Id: I3da770d2b9d39507ae06eb88550d89748ac170ed
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>